기존 수정 사항 업데이트

+++ 14-java/src/day09/practice01.java
@@ -16,7 +16,7 @@ public class practice01 {
 
         try {
             System.out.printf("The greatest common divisor of %d and %d = %d\n", n1, n2, getGcd(n1, n2));
-            System.out.printf("The least common multiple of %d and %d = %d\n", n1, n2, n1 * n2 / getGcd(n1, n2));
+            System.out.printf("The least common multiple of %d and %d = %d\n", n1, n2, (long)n1 * n2 / getGcd(n1, n2));
         } catch (IllegalArgumentException e) {
             System.err.println("ERROR: " + e.getMessage());
         } catch (Exception e) {
